requirements.txt
Flask==3.0.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
joblib==1.3.2
tifffile==2023.7.10
Werkzeug==3.0.1